templates/roadmap/aiml/learning/02-language-comparison/python/requirements.txt
fastapi
uvicorn[standard]==0.23.2
sqlalchemy
bcrypt
pyjwt
pydantic
passlib[bcrypt]